rewrote res as a double instead of an array, started mesh_conv and r_optimization tests

+++ include/num_methods.h
@@ -10,7 +10,7 @@ Defined in the context of laplace2d
 typedef struct s_parameters{
   int Ntype;
   double r;
-  int max_iter;
+  long int max_iter;
   int qtimes;
   int save_i_c; // Save initial condition
   int save_last_only;
